Create Contract-code
Entrepreneurs looking to launch a new cryptocurrency can do it through an initial coin offering (ICO), a variation on an initial public offering (IPO).
There is some government regulation of ICOs currently, and anyone can launch one, provided they get the technology put in place.
How does one put the technology in place? Create a white paper or other document outlining the system, make a website or app describing how it works, and seek funding.
Advertising is key since there are so many competing coins on the market, so figuring out how to appeal to the target demo is crucial.
Not looking to launch a new coin, but rather, to invest in a new coin? Make sure to do thorough research, as there are a number of scams.